correlaid-analytics/requirements.txt
tweepy
facebook-sdk
mailchimp3
pymysql